[arcane,core+std] Déplace les classes gérant les informations des fichiers MSH dans 'MshMeshGenerationInfo'.
