+++ arcane/src/arcane/core/MshMeshGenerationInfo.cc
@@ -22,6 +22,15 @@ namespace Arcane::impl
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
+MshMeshGenerationInfo::
+MshMeshGenerationInfo(IMesh* mesh)
+: m_mesh(mesh)
+{
+}
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
 } // namespace Arcane::impl
 
 /*---------------------------------------------------------------------------*/
